move upload step to reporter
