+++ src/index.ts
@@ -2,9 +2,8 @@ import Module from 'module'
 import * as path from 'path'
 import * as fs from 'fs'
 import { getInput, setFailed } from '@actions/core'
-import { getOctokit, context } from '@actions/github'
-import { reporter } from './reporter'
-import { CompilerOptions, Diagnostic, DiagnosticCategory, ParsedCommandLine } from "typescript"
+import { reporter, uploadAnnotations } from './reporter'
+import { CompilerOptions, Diagnostic, ParsedCommandLine } from "typescript"
 
 type TS = typeof import('typescript')
 
@@ -102,52 +101,8 @@ const performCompilation = (ts: TS, config:ParsedCommandLine) => {
   }
   const diagnostics = ts.sortAndDeduplicateDiagnostics(all)
 
-  try {
-
-    const repoToken = getInput('repo_token', { required: false })
-    if (repoToken) {
-      const octokit = getOctokit(repoToken)
-      const pullRequest = context.payload.pull_request
-      let ref
-      if (pullRequest) {
-        ref = pullRequest.head.sha
-      } else {
-        ref = context.sha
-      }
-      const owner = context.repo.owner
-      const repo = context.repo.repo
-  
-      // The GitHub API requires that annotations are submitted in batches of 50 elements maximum
-      const batchedAnnotations = batch(50, diagnostics.slice())
-      for (const batch of batchedAnnotations) {
-        const annotations = batch.map(diagnostic => {
-          return {
-            path: diagnostic.file,
-            start_line: diagnostic.start,
-            end_line: diagnostic.start,
-            annotation_level: getAnnotationLevel(diagnostic),
-            message: ts.flattenDiagnosticMessageText(diagnostic.messageText, '\n'),
-          }
-        })
-  
-        octokit.checks.update({
-          owner,
-          repo,
-          check_run_id: context.runId,
-          output: {
-            annotations
-          }
-        }).catch(err => {
-          console.log("upload fetch err", err)
-        })
-      }
-    }
-  } catch (error) {
-    console.log("error uploading annotations", error)
-  }
-    
+  uploadAnnotations(diagnostics.slice())
   diagnostics.forEach(diagnostic => report(diagnostic))
-
   return all.length
 }
 
@@ -164,31 +119,6 @@ const loadTS = (projectPath:string):TS => {
     const ts = require('typescript')
     console.log(`Failed to find project specific typescript, falling back to bundled typescript@${ts.version}`);
     return ts
-  }
-}
-
-function batch<T>(size: number, inputs: T[]){
-  return inputs.reduce((batches, input) => {
-    const current = batches[batches.length - 1]
-
-    current.push(input)
-
-    if (current.length === size) {
-      batches.push([])
-    }
-
-    return batches
-  }, [[]] as T[][])
-}
-
-const getAnnotationLevel = (diagnostic: Diagnostic) => {
-  switch (diagnostic.category) {
-    case DiagnosticCategory.Error:
-      return "failure"
-    case DiagnosticCategory.Warning:
-      return "warning"
-    default:
-      return "notice"
   }
 }
